doit.sh: port to Mac OS X
* add #!/bin/bash
* detect sha checksum tool

+++ doit.sh
@@ -1,3 +1,23 @@
+#!/bin/bash
+
+#
+# Attempt to detect installed sha1 tool.
+# sha1sum is likely on Linux, shasum is likely on Mac OS X.
+#
+for hasher in sha1sum shasum
+do
+	if [ "$(which ${hasher})" != "" ]
+	then
+		HASHER=${hasher}
+		break
+	fi
+done
+if [ ${HASHER} = "" ]
+then
+	echo "Could not find sha1sum tool. Try installing coreutils."
+	exit
+fi
+
 git clean -f -x -d -e doit.sh
 export CROSS_COMPILE=arm-none-eabi-
 
@@ -13,7 +33,7 @@ do
 (
 	make ${i}_config
 	make
-	sha1sum u-boot.img u-boot MLO spl/u-boot-spl.bin  > u-boot.sha1
+	${HASHER} u-boot.img u-boot MLO spl/u-boot-spl.bin  > u-boot.sha1
 	mkdir -p build/$i/spl
 	cp MLO build/$i/
 	cp u-boot build/$i/
